feature/add-movie-rating-data: Rename movie states
