--- data/src/main/java/com/london/data/remote/model/details/RatingRemote.kt
@@ -1,6 +1,0 @@
-package com.london.data.remote.model.details
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class RatingRemote(val value: Double)
