+++ data/src/main/java/com/london/data/remote/model/details/rating/RatingRemoteBody.kt
@@ -1,0 +1,6 @@
+package com.london.data.remote.model.details.rating
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class RatingRemoteBody(val value: Double)
