+++ data/src/main/java/com/london/data/remote/source/details/movie/MovieDetailsRemoteDataSource.kt
@@ -2,8 +2,8 @@ package com.london.data.remote.source.details.movie
 
 import com.london.data.remote.model.ApiResponse
 import com.london.data.remote.model.details.movie.model.moviecast.MovieCastResponse
+import com.london.data.remote.model.details.movie.model.moviedetails.MovieAccountStatesResponse
 import com.london.data.remote.model.details.movie.model.moviedetails.MovieDetailsResponse
-import com.london.data.remote.model.details.movie.model.moviedetails.MovieStatesResponse
 import com.london.data.remote.model.details.movie.model.movieimages.MovieImagesResponse
 import com.london.data.remote.model.search.model.SearchMovieRemote
 
@@ -13,9 +13,9 @@ interface MovieDetailsRemoteDataSource {
     suspend fun getSimilarMovies(movieId: Int): Result<ApiResponse<SearchMovieRemote>>
     suspend fun getMovieCast(movieId: Int): Result<MovieCastResponse>
     suspend fun getMovieImages(movieId: Int): Result<MovieImagesResponse>
-    suspend fun getMovieStates(
+    suspend fun getMovieAccountStates(
         movieId: Int,
         guestSessionId: String?,
         userSessionId: String?
-    ) : Result<MovieStatesResponse>
+    ) : Result<MovieAccountStatesResponse>
 }
