+++ data/src/main/java/com/london/data/remote/source/details/movie/rating/AddMovieRatingRemoteDataSourceImpl.kt
@@ -1,7 +1,7 @@
-package com.london.data.remote.source.details.movie
+package com.london.data.remote.source.details.movie.rating
 
-import com.london.data.remote.model.details.RatingRemote
-import com.london.data.remote.model.details.RatingResponse
+import com.london.data.remote.model.details.rating.RatingRemoteBody
+import com.london.data.remote.model.details.rating.RatingResponse
 import com.london.data.remote.service.details.movie.AddMovieRatingApiService
 import com.london.data.remote.source.base.BaseRemoteDatasource
 import javax.inject.Inject
@@ -17,7 +17,7 @@ class AddMovieRatingRemoteDataSourceImpl @Inject constructor(
     ): Result<RatingResponse> = callApiWithRetry(apiCall = {
         addMovieRatingApiService.addMovieRating(
             movieId = movieId,
-            ratingRequest = RatingRemote(rating),
+            ratingRequest = RatingRemoteBody(rating),
             guestSessionId = guestSessionId,
             userSessionId = userSessionId
         )
